+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [3.16.3] - 2025-01-21
+### Fixed
+- Bug with async boolean expressions (when an operator takes an async value as
+  a singleton parameter).
+
 ## [3.16.2] - 2025-01-16
 ### Fixed
 - Bug with toString when userInvocationTable passed.